B2C Marketing automation report 2017
The concept of marketing automation has evolved and refined over the years since the early 2000s to have now become more intuitive and relevant. The strategy of batching up all your leads and blasting the same SMS or email out to each one of them, was anything-but-strategic, disjointed and was potentially leading to a loss of consumer interest in advertising messages. A few years later the concept of “personalization” became critical, thanks to the mine of customer information that digital channels across the web, mobile and social media provided to the marketers. Advertising became more targeted, meaningful and effective, because the right message was being sent to the right target. Interacting with customer has now become intuitive, highly segmented and behaviour-triggered thanks to rapid advances in marketing automation technology that enables marketers to literally follow the customer around wherever he goes!